Move controller_stopper to ur_robot_driver
Since a standalone package with that name was declined during the release
process, we decided to move the controller_stopper package over to the driver.
