--- controller_stopper/CHANGELOG.rst
@@ -1,15 +1,0 @@
-^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-Changelog for package controller_stopper
-^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-
-2.1.0 (2022-12-08)
-------------------
-* Drop old C++ compiler flags (`#577 <https://github.com/UniversalRobots/Universal_Robots_ROS_Driver/issues/577>`_)
-  Co-authored-by: Jochen Sprickerhof <git@jochen.sprickerhof.de>
-* Remove boilerplate comments from package.xml files (`#528 <https://github.com/UniversalRobots/Universal_Robots_ROS_Driver/issues/528>`_)
-* Contributors: Felix Exner, Michael Görner
-
-2.0.0 (2021-09-07)
-------------------
-* Initial release
-* Contributors: Felix Exner, Felix Mauch, G.A. vd. Hoorn, Tristan Schnell
