--- controller_stopper/package.xml
@@ -1,22 +1,0 @@
-<?xml version="1.0"?>
-<package format="2">
-  <name>controller_stopper</name>
-  <version>2.1.0</version>
-  <description>A small helper node that stops and restarts ROS controllers based on a boolean status topic.</description>
-
-  <maintainer email="exner@fzi.de">Felix Exner</maintainer>
-
-  <license>Apache-2.0</license>
-
-  <buildtool_depend>catkin</buildtool_depend>
-  <build_depend>controller_manager_msgs</build_depend>
-  <build_depend>roscpp</build_depend>
-  <build_depend>std_msgs</build_depend>
-  <build_export_depend>controller_manager_msgs</build_export_depend>
-  <build_export_depend>roscpp</build_export_depend>
-  <build_export_depend>std_msgs</build_export_depend>
-  <exec_depend>controller_manager_msgs</exec_depend>
-  <exec_depend>roscpp</exec_depend>
-  <exec_depend>std_msgs</exec_depend>
-
-</package>
