+++ ur_robot_driver/launch/ur_control.launch
@@ -81,7 +81,7 @@
   <node name="ros_control_stopped_spawner" pkg="controller_manager" type="spawner" respawn="false"
     output="screen" args="--stopped $(arg stopped_controllers)" unless="$(eval arg('stopped_controllers') == '')"/>
 
-  <node name="controller_stopper" pkg="controller_stopper" type="node" respawn="false" output="screen">
+  <node name="controller_stopper" pkg="ur_robot_driver" type="controller_stopper_node" respawn="false" output="screen">
     <remap from="robot_running" to="ur_hardware_interface/robot_program_running"/>
     <rosparam param="consistent_controllers">
        - "joint_state_controller"
